+++ haystack/preview/components/generators/openai/_helpers.py
@@ -1,10 +1,15 @@
 import logging
 
+from haystack.preview.lazy_imports import LazyImport
+
+with LazyImport("Run 'pip install tiktoken'") as tiktoken_import:
+    import tiktoken
+
 
 logger = logging.getLogger(__name__)
 
 
-def enforce_token_limit(prompt: str, tokenizer, max_tokens_limit: int) -> str:
+def enforce_token_limit(prompt: str, tokenizer: "tiktoken.Encoding", max_tokens_limit: int) -> str:
     """
     Ensure that the length of the prompt is within the max tokens limit of the model.
     If needed, truncate the prompt text so that it fits within the limit.
@@ -14,15 +19,15 @@ def enforce_token_limit(prompt: str, tokenizer, max_tokens_limit: int) -> str:
     :param max_tokens_limit: The max tokens limit of the model.
     :return: The prompt text that fits within the max tokens limit of the model.
     """
+    tiktoken_import.check()
     tokens = tokenizer.encode(prompt)
     tokens_count = len(tokens)
     if tokens_count > max_tokens_limit:
         logger.warning(
-            "The prompt has been truncated from %s tokens to %s tokens so that the prompt fits within the max token "
-            "limit. Reduce the length of the prompt to prevent it from being cut off.",
+            "The prompt has been truncated from %s tokens to %s tokens to fit within the max token limit. "
+            "Reduce the length of the prompt to prevent it from being cut off.",
             tokens_count,
             max_tokens_limit,
         )
-        tokenized_payload = tokenizer.encode(prompt)
-        prompt = tokenizer.decode(tokenized_payload[:max_tokens_limit])
+        prompt = tokenizer.decode(tokens[:max_tokens_limit])
     return prompt
